requirements.txt
selenium==4.23.1
fpdf==1.7.2
PyAutoGUI==0.9.54
pandas==2.2.2
pillow==10.4.0